[gentoo-certbot] remove package
As this is a CLI tool and not a service that needs constant running,
decision has been made to remove this.

Main purpose of this project is to provide CPU optimized binaries
which offers performance improvements. This improvements can only be
noticible when a sevice is ran for long period and not just for few
seconds (like in CLI)